solved relop

+++ HW1/tokens.hpp
@@ -42,6 +42,6 @@ enum tokentype {
 
 extern int yylineno;
 extern char* yytext;
-extern int yyleng;
+extern size_t yyleng;
 extern int yylex();
 #endif /* TOKENS_HPP_ */
